Move Task analyzer to specific package
